+++ optimus/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.12.9"
+__version__ = "0.12.10"
 
 
 def safe_commit() -> None:
